Rename Global gateway to Payeezy gateway
