--- src/Message/GlobalAuthorizeRequest.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Omnipay\FirstData\Message;
-
-class GlobalAuthorizeRequest extends GlobalAbstractRequest
-{
-    protected $action = self::TRAN_PREAUTH;
-}
