--- src/Message/GlobalPurchaseRequest.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Omnipay\FirstData\Message;
-
-class GlobalPurchaseRequest extends GlobalAbstractRequest
-{
-    protected $action = self::TRAN_PURCHASE;
-}
