+++ src/Message/PayeezyCaptureRequest.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * First Data Payeezy Capture Request
+ */
+
+namespace Omnipay\FirstData\Message;
+
+/**
+ * First Data Payeezy Capture Request
+ */
+class PayeezyCaptureRequest extends PayeezyRefundRequest
+{
+    protected $action = self::TRAN_TAGGEDPREAUTHCOMPLETE;
+}
